requirements.txt
openai==0.28.1
python-dotenv
langchain
langchain-text-splitters
sentence-transformers>=2.7.0
torch>=2.2
numpy
faiss-cpu
transformers
requests
beautifulsoup4
pypdf2
pdfminer.six